Move tagging functions to utils file
